UpdateGetPage.py
I decided to use API because it will be easier to read codes in file. Using that command
dir_re = re.compile(r'<a.*?href="(.+?)".*?class="js-navigation-open".*?>(?:<span.*?<\/span>)?(.+?)(\..*)?<\/a>') will cause an error. Because Ghdata has so many contributors. One of them could say "class="js-navigation-open" class name could be really long and lets make it "class = "js-nav"" When they make this change, We will not able to use our code, it wil not work.